requirements.txt
pycoingecko
pandas
xmltodict
pillow
moviepy
numpy
matplotlib
opencv-python
opencv-contrib-python